huge progress
changed list array to use object literals for tasks instead of just a single string, no need for a separate completed array and functions that handle it, im embarrassed i didnt think of this sooner. added some more styling and task sorting button on the bottom of the app. all thats left now is to fix the stupid theme switching and make website responsive